Feature/approve project blockchain integration

+++ app/services/blockchain/contracts/offering_service.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from app.services.blockchain.base_contract_service import BaseContractService
 from app.services.blockchain.deployment import DeploymentReader
 from app.core.config import settings
@@ -7,19 +8,26 @@ import time
 class OfferingService(BaseContractService):
     contract_name = "Offering"
 
-    def deploy(self, dpf_token: str, soft_cap: int, hard_cap: int, token_price: int, deadline_seconds: int) -> str:
+    def deploy(
+        self,
+        dpf_token: str,
+        soft_cap: int,
+        hard_cap: int,
+        token_price: Decimal,
+        deadline_seconds: int,
+    ) -> str:
         addresses = DeploymentReader.get_addresses()
         deadline = int(time.time()) + deadline_seconds
 
         address = self.client.deploy_contract(
             self.contract_name,
-            addresses["usdc_address"],   # ← antes settings.USDC_ADDRESS
+            addresses["usdc_address"],
             dpf_token,
-            soft_cap,
-            hard_cap,
-            token_price,
+            int(soft_cap),
+            int(hard_cap),
+            int(token_price),
             deadline,
-            settings.TREASURY_ADDRESS,   # ← este se queda en settings
+            settings.TREASURY_ADDRESS,
         )
 
         self._contract = self.client.get_contract(self.contract_name, address)
